requirements.txt
aioredis==1.2.0
amqp==2.5.0
argon2-cffi==19.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
awesome-slugify==1.6.5
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0rc3
certifi==2019.6.16
cffi==1.12.3
channels==2.1.7
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.9
django-allauth==0.39.1
django-appconf==1.0.3
django-celery-beat==1.5.0
django-celery-email==2.0.1
django-compressor==2.3
django-contrib-comments==1.9.1
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-formtools==2.1
django-import-export==1.2.0
django-markdownx==2.0.28
django-mdeditor==0.1.16
django-redis==4.10.0
django-taggit==1.1.0
django-test-plus==1.3.1
django-timezone-field==3.0
elasticsearch==2.0.0
et-xmlfile==1.0.1
future==0.17.1
hiredis==1.0.0
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
kombu==4.6.3
Markdown==3.1.1
msgpack==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.2
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
PyHamcrest==1.9.0
python-crontab==2.3.7
python-dateutil==2.8.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
rcssmin==1.0.6
redis==3.2.1
regex==2019.6.8
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.1.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tablib==0.13.0
text-unidecode==1.2
urllib3==1.24.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
